Modificaciones en Ambitos de los Recursos para que se validen de forma correcta los campos obligatorios

+++ src/main/java/com/dbcom/app/model/dao/AmbitoRecursoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.dbcom.app.model.entity.AmbitoRecurso;
 
 @Repository
-public interface AmbitoRecursoRepository extends JpaRepository<AmbitoRecurso, Short>{
+public interface AmbitoRecursoRepository extends JpaRepository<AmbitoRecurso, Long>{
 
 }
